pyproject.toml
[project]
name = "langchain-sql-agent-demo"
version = "0.1.0"
description = "AMP-style SQL task orchestrator with PostgreSQL CRUD and plan/confirm workflow"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "langchain>=0.2.16",
    "langchain-community>=0.2.16",
    "langchain-core>=0.2.43",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "pydantic>=2.7",
    "flask>=3.0",
    "datasets>=4.5.0",
    "huggingface-hub>=1.4.1",
    "gdown>=5.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[project.scripts]
sql-agent = "sql_agent_demo.cli:main"
sql-agent-api = "sql_agent_demo.interfaces.api:main"
sql-agent-demo = "sql_agent_demo.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sql_agent_demo"]

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=7.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: pure function and module tests with fixed inputs and no external services",
    "db: database adapter and schema initialization tests",
    "api: Flask /run contract tests",
    "integration: multi-component backend integration tests",
    "smoke: browser or end-to-end smoke tests",
    "postgres: PostgreSQL-path verification",
    "sqlite_fallback: SQLite fallback verification",
    "real_model: tests that require a live model proxy",
    "fake_model: deterministic tests that use fake models",
]
